Spitted the redux store tests into separate test files (only moved code, no changes)
