+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/model/immutable/StateCompareTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 733a0d4de7e63b4418eddcc4bda8e891
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
